enhance lang switcher
