+++ src/i18n.ts
@@ -1,6 +1,6 @@
-import { register, init, getLocaleFromNavigator } from 'svelte-i18n';
+import { register, init } from 'svelte-i18n';
 
 register('en', () => import('./locales/en.json'));
 register('ar', () => import('./locales/ar.json'));
 
-init({ initialLocale: getLocaleFromNavigator(), fallbackLocale: 'en' });
+init({ initialLocale: 'en', fallbackLocale: 'en' });
